pyproject.toml
[project]
name = "mysystemdown"
version = "0.1.0"
description = "MySystemDown is a health check system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-tenants>=3.10.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
]
